+++ crc.h
@@ -3,6 +3,6 @@
 
 #include <stdint.h>
  
-uint32_t crc32(const uint8_t* buffer, int length, uint32_t startValue = 0);
+uint32_t crc32(const uint8_t* buffer, size_t length, uint32_t startValue = 0);
 
 #endif
